Package structure update
