+++ src/apify/events/__init__.py
@@ -1,0 +1,5 @@
+from crawlee.events import EventManager, LocalEventManager
+
+from ._apify_event_manager import ApifyEventManager
+
+__all__ = ['ApifyEventManager', 'EventManager', 'LocalEventManager']
