+++ src/apify/request_loaders/__init__.py
@@ -1,0 +1,18 @@
+from crawlee.request_loaders import (
+    RequestList,
+    RequestLoader,
+    RequestManager,
+    RequestManagerTandem,
+    SitemapRequestLoader,
+)
+
+from ._apify_request_list import ApifyRequestList
+
+__all__ = [
+    'ApifyRequestList',
+    'RequestList',
+    'RequestLoader',
+    'RequestManager',
+    'RequestManagerTandem',
+    'SitemapRequestLoader',
+]
