+++ src/apify/storages/__init__.py
@@ -1,5 +1,3 @@
 from crawlee.storages import Dataset, KeyValueStore, RequestQueue
 
-from ._request_list import RequestList
-
-__all__ = ['Dataset', 'KeyValueStore', 'RequestList', 'RequestQueue']
+__all__ = ['Dataset', 'KeyValueStore', 'RequestQueue']
